+++ common/src/main/java/io/github/mortuusars/scholar/Scholar.java
@@ -9,6 +9,7 @@ import net.minecraft.core.component.DataComponentType;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.tags.TagKey;
@@ -16,6 +17,7 @@ import net.minecraft.util.ExtraCodecs;
 import net.minecraft.util.Unit;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.inventory.MenuType;
+import net.minecraft.world.level.storage.loot.LootTable;
 import org.slf4j.Logger;
 
 import java.util.function.Supplier;
@@ -78,5 +80,9 @@ public class Scholar {
         public static class EntityTypes {
             public static final TagKey<EntityType<?>> LITERATE = TagKey.create(Registries.ENTITY_TYPE, resource("literate"));
         }
+    }
+
+    public static class LootTables {
+        public static final ResourceKey<LootTable> LITERATE_MOB_BOOK = ResourceKey.create(Registries.LOOT_TABLE, resource("entities/literate_mob_book"));
     }
 }