+++ common/src/main/java/io/github/mortuusars/scholar/world/entity/LiterateMobs.java
@@ -5,19 +5,18 @@ import io.github.mortuusars.scholar.Scholar;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.DifficultyInstance;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.EquipmentTable;
 import net.minecraft.world.entity.Mob;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
+
+import java.util.Map;
 
 public class LiterateMobs {
     public static void populateEquipmentSlots(RandomSource random, DifficultyInstance difficulty, Mob mob) {
         if (mob.getType().is(Scholar.Tags.EntityTypes.LITERATE)
               && random.nextDouble() < Config.Common.LITERATE_MOBS_BOOK_SPAWN_CHANCE.get()) {
-            mob.setItemSlot(EquipmentSlot.MAINHAND, new ItemStack(Items.BOOK));
-            // This can potentially make drop chances of other items higher that they should be,
-            // if code down the line will override the main hand slot without changing drop chances.
-            // But I don't know of an easy solution to this, without mixing in into each mob individually.
-            mob.setDropChance(EquipmentSlot.MAINHAND, (float) Config.Common.LITERATE_MOBS_BOOK_DROP_CHANCE.getAsDouble());
+            float dropChance = (float) Config.Common.LITERATE_MOBS_BOOK_DROP_CHANCE.getAsDouble();
+            mob.equip(new EquipmentTable(Scholar.LootTables.LITERATE_MOB_BOOK,
+                  Map.of(EquipmentSlot.MAINHAND, dropChance, EquipmentSlot.OFFHAND, dropChance)));
         }
     }
 }